fix(LLM): Deeplinks Refacto update
Separated Styles and Performance from the Deeplink Provider. Created two new providers in index.tsx of the app. They remain here for now as they are sufficiently short
